Add "--output file" command-line option
Default is to send output to stdout.  Otherwise to the given file

Also reorganize some of the modules
