+++ centrinel.cabal
@@ -26,6 +26,9 @@ tested-with:         GHC == 7.10.*, GHC == 8.0.*, GHC == 8.2.*
 
 library
   exposed-modules:     Centrinel,
+                       Centrinel.InRepl,
+                       Centrinel.Types,
+                       Centrinel.Report,
                        Centrinel.Warning,
                        Centrinel.Trav,
                        Centrinel.Region.Region,
