+++ main/Main.hs
@@ -3,6 +3,7 @@ module Main where
 import Data.Monoid 
 import Options.Applicative
 import qualified Centrinel.Main
+import qualified Centrinel.Report
 
 main :: IO ()
 main = execParser theParser >>= Centrinel.Main.main
@@ -10,8 +11,10 @@ main = execParser theParser >>= Centrinel.Main.main
 theParser :: ParserInfo Centrinel.Main.CentrinelCmd
 theParser =
   let
-    commonOptions = Centrinel.Main.CentrinelOptions <$> useCCOption
+    commonOptions :: Parser Centrinel.Main.CentrinelOptions
+    commonOptions = Centrinel.Main.CentrinelOptions <$> useCCOption <*> outputMethodOptions
 
+    useCCOption :: Parser (Maybe FilePath)
     useCCOption = optional $ strOption (long "use-cc" <> metavar "CC" <> help "Use the given C compiler for preprocessing (Default: CC environment value or 'cc' if unset)")
 
     runProject :: Parser (Centrinel.Main.CentrinelOptions -> Centrinel.Main.CentrinelCmd)
@@ -20,3 +23,17 @@ theParser =
     runOne = flip Centrinel.Main.RunOneCentrinelCmd <$> some (strArgument $ metavar "-- CFLAGS... CFILE" <> help "Run the analysis on a single C input file with the given flags")
     parser = helper <*> commonOptions <**> (runProject <|> runOne)
   in info parser (fullDesc <> (progDesc "centrinel is a C static analyzer"))
+
+
+outputMethodOptions :: Parser Centrinel.Report.OutputMethod
+outputMethodOptions = Centrinel.Report.OutputMethod <$> outputDestinationOption <*> outputFormatOption
+  where
+    outputDestinationOption :: Parser Centrinel.Report.OutputDestination
+    outputDestinationOption =
+      (Centrinel.Report.FilePathOutputDestination <$> strOption (long "output" <> help "destination file for the output"
+                                                               <> metavar "FILE" ))
+      <|> pure Centrinel.Report.StdOutOutputDestination 
+
+    outputFormatOption :: Parser Centrinel.Report.OutputFormat
+    outputFormatOption = pure Centrinel.Report.PlainTextOutputFormat
+
